[CSL-1279] Merge Pos.Arbitrary.Ssc and ...GodTossing
